V2 Queries split and deprecation notice
Making readme changes per PR comments , added deprecation notice
Also split out V2 queries into its own file.
